set up chore list
